initial commit
A programme that checks for solutions to the Strand Puzzle. 

Imagine a street with numbers ascending in order: 1, 2, 3, 4, etc.

A man stands at his front door and looks to his left and right.

He realises that the total of the numbers descending below his house number down to 1 equals the total of the ascending numbers up to the end of the street. 

How long is his street? And which house number is his?

Well there are lots of examples, but the smallest street possible is eight houses long, and the man stands at house number 6.

To illustrate (6 is not included as the man is standing outside house number 6 looking left and right):

1 + 2 + 3 + 4 + 5 = 15
7 + 8 = 15

The possible solutions get big very quickly.